requirements.txt
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
google-api-core==1.26.1
google-auth==1.28.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-secret-manager==2.3.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.36.1
grpcio-reflection==1.36.1
grpcio-tools==1.36.1
idna==2.10
libcst==0.3.17
mypy-extensions==0.4.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
